+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `batch_submit` task - [#41](https://github.com/PrefectHQ/prefect-aws/issues/41)
+- `MinIOCredentials` block - [#46](https://github.com/PrefectHQ/prefect-aws/pull/46)
 
 ### Changed